feat: move ChannelFeedState, SplitBuilder, ForwardingPainter, and Toast classes to commons
Move KMP-compatible UI utilities and data classes to commons module:

- ChannelFeedState: sealed class for channel feed states (uses commons Channel)
- SplitBuilder/SplitItem: generic split percentage builder (compose runtime only)
- ForwardingPainter/ForwardingDrawInfo: painter wrapper utility (compose UI only)
- ToastMsg/StringToastMsg/ActionableStringToastMsg: toast message hierarchy
- ThrowableToastMsg/ThrowableToastMsg2: error toast messages

All originals replaced with typealiases for backwards compatibility.
Consumers updated to import from commons where needed.
