+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/ui/components/toasts/StringToastMsg.kt
@@ -18,10 +18,18 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.components.toasts
+package com.vitorpamplona.amethyst.commons.ui.components.toasts
 
-import com.vitorpamplona.amethyst.commons.ui.components.toasts.ActionableStringToastMsg as CommonsActionableStringToastMsg
-import com.vitorpamplona.amethyst.commons.ui.components.toasts.StringToastMsg as CommonsStringToastMsg
+import androidx.compose.runtime.Immutable
 
-typealias StringToastMsg = CommonsStringToastMsg
-typealias ActionableStringToastMsg = CommonsActionableStringToastMsg
+@Immutable
+class StringToastMsg(
+    val title: String,
+    val msg: String,
+) : ToastMsg()
+
+class ActionableStringToastMsg(
+    val title: String,
+    val msg: String,
+    val action: () -> Unit,
+) : ToastMsg()
